structure change for docker image run
