--- Frontend/.env.local.example
@@ -1,2 +1,0 @@
-# Backend base URL (no trailing slash)
-NEXT_PUBLIC_API_BASE_URL=http://127.0.0.1:8000
