+++ Frontend/package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "axios": "^1.13.5",
     "next": "14.2.18",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
